move navs out from `admin`
