+++ gwcli/tree/license/license.go
