+++ gwcli/tree/users/self/interactive.go
