+++ gwcli/tree/users/self/myinfo.go
