+++ gwcli/tree/users/self/myinfo_noci_test.go
