+++ gwcli/tree/users/self/self_noci_test.go
