+++ gwcli/tree/users/toggle_admin.go
